% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_loader_TMP.R
\name{decision_loader}
\alias{decision_loader}
\title{Using imported population and control parameters, run a population viability analysis (PVA) on the target species.}
\usage{
decision_loader(csv_input = NULL, list_input = NULL)
}
\arguments{
\item{input}{List of input parameters, usually created with the load_pva_parameters() function.

decision_setup(input = load_pva_parameters("parameters.csv", scen_names = c("Scen1", "Scen2"))}
}
\description{
Using imported population and control parameters, run a population viability analysis (PVA) on the target species.
}
