% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_setup.R
\name{decision_setup}
\alias{decision_setup}
\title{Using imported population and control parameters, run a population viability analysis (PVA) on the target species.}
\usage{
decision_setup(
  params,
  scen_names,
  list = T,
  csv = F,
  csv_path = NULL,
  gui = T,
  selected_params = NULL
)
}
\arguments{
\item{params}{List of params parameters, usually created with the load_pva_parameters() function.}

\item{scen_names}{Character vector of scenario names.}

\item{list}{(Optional) Boolean; should the output be a named list? See below for list format. Either list or csv must be TRUE.}

\item{csv}{(Optional) Boolean; should the output be a CSV to be re-loaded into the session? If csv=TRUE, csv_path must be provided.}

\item{csv_path}{(Optional) If csv=TRUE, where should the decision() template be saved?}

\item{gui}{(Optional) If TRUE, parameters will be selected by the user via a pop-up menu. Otherwise, selected_params must be provided.}

\item{selected_params}{(Optional) If gui = FALSE, the user can provide valid named parameters in a character vector here. Alternatively, "all" will create a decision template including all control parameters. Must be provided if gui = FALSE.

Outputs either a nested named list (if list = TRUE) or a csv (if csv = TRUE and csv_path provided).
 -  List output: a nested named list where the outermost list contains names of scenarios, and within each scenario is a named list of parameters to modify for decision-making.
 -  CSV output: a CSV file where the first column indicates scenario names, and subsequent columns indicate control parameters to be varied between scenarios.
When setting up params files for the decision, users should leave blank spaces where the parameters should be the same as the params parameters. Only those values that are filled in will be modified and compared with decision().}
}
\description{
Using imported population and control parameters, run a population viability analysis (PVA) on the target species.
}
