% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_sensitivity.R
\name{ggplot_sensitivity}
\alias{ggplot_sensitivity}
\title{Rank proposed PVA decision scenarios when biological parameters are increased or decreased by some percentage.}
\usage{
ggplot_sensitivity(df, ylimit = NaN)
}
\arguments{
\item{df}{Data frame containing data for plotting.}

\item{ylimit}{(Optional) The upper limit of the y-axis.}
}
\description{
Rank proposed PVA decision scenarios when biological parameters are increased or decreased by some percentage.
}
