% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pva.R
\name{plot_pva}
\alias{plot_pva}
\title{Plot the outputs of a PVA simulation.}
\usage{
plot_pva(pva, set.ymax = NULL, quiet = FALSE)
}
\arguments{
\item{pva}{List of outputs created by running the PVA() function.}
}
\value{
List of
* A list of PVA outputs, including calculated parameters:
 - `phie`: unfished eggs per recruit,
 - `R.A` and `R.B`: alpha and beta paramters for a Beverton-Holt recruitment function,
 - `Rinit`: initial recruitment,
 - `amat`: age at maturity to differentiate adults and sub-adults,
 - `A.s`: maximum survival in each stanza, and
 - `B.s`: carrying capacity parameter for each stanza)

and PVA-generated calculations for each iteration of the simulation
 - `N1`: initial population size per simulation,
}
\description{
Plot the outputs of a PVA simulation.
}
