% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pva_template.R
\name{pva_template}
\alias{pva_template}
\title{Save a .csv template of parameters. Once filled in, the csv template can be loaded using load_pva_parameters().}
\usage{
pva_template(file)
}
\arguments{
\item{filepath}{File path where the parameter template should be saved.}
}
\value{
Nothing - a parameter template will be saved at the `file` location provided.
}
\description{
Save a .csv template of parameters. Once filled in, the csv template can be loaded using load_pva_parameters().
}
