% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_uncertainty.R
\name{rank_uncertainty}
\alias{rank_uncertainty}
\title{Rank proposed PVA decision scenarios when biological parameters are increased or decreased by some percentage.}
\usage{
rank_uncertainty(
  input,
  percent = 0.15,
  decision_csv = NULL,
  decision_list = NULL,
  parallel = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{percent}{The percentage by which biological parameters ( ) should be modified. Defaults to 15%, such that biological parameters are multiplied by 115% and 85% for upper and lower estimates, respectively.}

\item{parallel}{(Optional) Set parallel = TRUE to run analyses using multiple cores.}

\item{decision}{The output from a decision() function call. Provides the basis for comparing rankings of control scenarios.}
}
\description{
Rank proposed PVA decision scenarios when biological parameters are increased or decreased by some percentage.
}
