% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision.R
\name{decision}
\alias{decision}
\title{Run multiple PVA simulations initialized with different control scenarios to compare their simulated costs and outcomes.}
\usage{
decision(
  input,
  decision_csv = NULL,
  decision_list = NULL,
  sens.pcent = NULL,
  direction = NULL,
  sens.params = NULL,
  run_parallel = F
)
}
\arguments{
\item{sens.pcent}{(Optional, invoked by the `rankUncertainty` function) For the sake of sensitivity analysis, how much should population parameters (i.e. \code{}, \code{}, \code{}, \code{}, \code{}, \code{}, \code{})}

\item{direction}{(Optional, invoked by the `rankUncertainty` function) Should biological parameters be increased or decreased by  `sens.percent`?}

\item{custom.inits}{(Optional, invoked by the `rankUncertainty` function) A vector containing the names of which parameters, if any, should differ from the values provided in \code{pva.params}.}
}
\value{

}
\description{
Run multiple PVA simulations initialized with different control scenarios to compare their simulated costs and outcomes.
}
