% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_sensitivity.R
\name{ggplot_sensitivity}
\alias{ggplot_sensitivity}
\title{Rank proposed PVA decision scenarios when biological parameters are increased or decreased by some percentage.}
\usage{
ggplot_sensitivity(df, ylimit = NaN)
}
\description{
Rank proposed PVA decision scenarios when biological parameters are increased or decreased by some percentage.
}
