% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize parameters for use in a \code{pva()} simulated run; called by the \code{pva()} function.}
\usage{
init(input, input.params = NULL, p.cent.trans = NULL)
}
\description{
Initialize parameters for use in a \code{pva()} simulated run; called by the \code{pva()} function.
}
