% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pva_parameters.R
\name{load_pva_parameters}
\alias{load_pva_parameters}
\title{Load a .csv template of parameters. A blank template that matches the parameters needed by `PVAInvasR` can be created with the pva_template() function.}
\usage{
load_pva_parameters(filepath)
}
\arguments{
\item{filepath}{The path to the filled-in .csv template of population and control parameters.}
}
\value{
pva_parameters A list of population and control parameters for use in other `PVAInvasR` functions.
}
\description{
Load a .csv template of parameters. A blank template that matches the parameters needed by `PVAInvasR` can be created with the pva_template() function.
}
\examples{
load_pva_parameters('~/Documents/PVA_Examples/filledin_parameter_template.csv')
}
