% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PVA.R
\name{PVA}
\alias{PVA}
\title{Using imported population and control parameters, run a population viability analysis (PVA) on the target species.}
\usage{
PVA(
  parameters,
  inits = NULL,
  custom.inits = NULL,
  sens.pcent = NULL,
  sens.params = NULL
)
}
\arguments{
\item{custom.inits}{(Optional) A vector containing the names of which parameters, if any, should differ from the values provided in \code{pva.params}. Should be a named list of po  Can be be outputs of the \code{init()} function from \code{PVAInvas}.}

\item{sens.pcent}{(Optional) For the sake of sensitivity analysis, how much should population parameters (i.e. \code{}, \code{}, \code{}, \code{}, \code{}, \code{}, \code{})}

\item{params}{A list of initialized population and control parameters to inform the PVA. Parameters should be provided in the form of a named list. We suggest filling in a parameter template, which can be created and loaded using the \code{pva_template()} and \code{load_pva_parameters()} functions.}
}
\value{
pva 
* A list of PVA outputs, including calculated parameters:
 - `phie`: unfished eggs per recruit,
 - `R.A` and `R.B`: alpha and beta paramters for a Beverton-Holt recruitment function,
 - `Rinit`: initial recruitment,
 - `amat`: age at maturity to differentiate adults and sub-adults,
 - `A.s`: maximum survival in each stanza, and
 - `B.s`: carrying capacity parameter for each stanza)

and PVA-generated calculations for each iteration of the simulation
 - `N1`: initial population size per simulation,

)
}
\description{
Using imported population and control parameters, run a population viability analysis (PVA) on the target species.
}
\examples{
# Run a simple PVA, no custom values or sensitivity testing.
PVA(pva.params = inputParameterList)
}
