% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pva.R
\name{PVA}
\alias{PVA}
\title{Using imported population and control parameters, run a population viability analysis (PVA) on the target species.}
\usage{
PVA(pva.params, custom.inits = NULL, sens.pcent = NULL, sens.params = NULL)
}
\arguments{
\item{pva.params}{A list of population and control parameters to inform the PVA. Parameters should be provided in the form of a named list. We suggest filling in a parameter template, which can be created and loaded using the `pva_template()` and `load_pva_parameters()` functions.}
}
\value{
pva A list of PVA outputs, including calculated parameters (`phie`:
 unfished eggs per recruit, `R.A` and `R.B`: alpha and beta paramters for a
 Beverton-Holt recruitment function, `Rinit`: initial recruitment,
`amat`: age at maturity to differentiate adults and sub-adults,
`A.s`: maximum survival in each stanza, and
`B.s`: carrying capacity parameter for each stanza)
and PVA-generated calculations for each iteration of the simulation
(`N1`: initial population size per simulation,

)
}
\description{
Using imported population and control parameters, run a population viability analysis (PVA) on the target species.
}
\examples{
load_pva_parameters('~/Documents/PVA_Examples/filledin_parameter_template.csv')
}
