% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pva_template.R
\name{pva_template}
\alias{pva_template}
\title{Save a .csv template of parameters. Once filled in, the csv template can be loaded using load_pva_parameters().}
\usage{
pva_template(file)
}
\arguments{
\item{filepath}{File path where the parameter template should be saved.}
}
\value{
Nothing - a parameter template will be saved at the file location provided.
Proportional to ascending slope of dome-shaped selectivity curve',
Proportional to length at maximum selectivity',
Rate of decline of dome-shaped selectivity (bounded between 0-1)',
}
\description{
Save a .csv template of parameters. Once filled in, the csv template can be loaded using load_pva_parameters().
}
